+++ rllib/algorithms/td_learning/tdl1.py
@@ -10,5 +10,5 @@ class TDL1(AbstractTDLearning):
 
     def _update(self, td_error, phi, next_phi, weight):
         weight_minus = torch.ones_like(weight) / weight
-        s = Bernoulli(torch.tensor(weight / (weight_minus + weight))).sample().float()
+        s = Bernoulli(torch.tensor(weight / (weight_minus + weight))).sample()
         self.theta += self.lr_theta * s @ (phi - self.gamma * next_phi)
